signing/stronghold: fix mnemonic probing logic
Because the snapshot is lazily loaded, calling store_mnemonic()
right after creating a StrongholdSigner instance with an existing
snapshot_path will allow one to still overwrite the mnemonic
stored in its vault, if we only look at the snapshot_loaded flag.

Try loading the snapshot before storing a mnemonic, but don't care
about whether it's succeeded or not to handle such case.