Docs - Edit with Style Guide and Implement Diataxis (#807)
* Deleted unnecessary files
Edited text in all sections except library examples which will be addressed in subsequent PRs
Refactored Specs table and header structure to be readable in wiki (see builder table).  @todo check if we can sort this out in CSS as this file is probably autogenerated.  Alternatively, see if we can tinker with the generator.
Added diataxis tags.
Updated iota-wiki-cli version.

* Fixed spacing

* Deleted package-lock.json and added it to gitignore

* Apply suggestions from code review from @charlesthompson3

Co-authored-by: charlesthompson3 <74603461+charlesthompson3@users.noreply.github.com>

* Update documentation/docs/contribute.md

Co-authored-by: Aleksei Korolev <nice.guy.alex.korolev@gmail.com>

Co-authored-by: charlesthompson3 <74603461+charlesthompson3@users.noreply.github.com>
Co-authored-by: Aleksei Korolev <nice.guy.alex.korolev@gmail.com>