Fix Workflows (#951)
* ci,core: fix rustc & clippy warnings & errors

CI workflows are slightly fixed and tweaked. Missing dependencies
for Ledger hardware support are installed. Some known security
vulnerabilities are false-positive and are ignored.

WASM compilation is switched from feature gating to target_family
gating. It's now possible to use --all-features when the target
isn't WASM. WASM targets still cannot use it, as some features
aren't supported on WASM.

* message_interface: mark panicking tests

They rely on the environment (.env file and a node) to run, which
isn't appropriate for unit tests. To make CIs pass, we mark them
with #[should_panic]. They should be changed to not rely on the
environment or be put under examples/ later.