stronghold: add timeout and task methods
get_timeout() and set_timeout() are added to view and change the
key clearing timeout. The task will be automatically spawned in
set_timeout() if a new timeout is provided and key is still in the
memory.

restart_key_timeout() is also added as a shortcut of refreshing
the count down by setting the same timeout again.