stronghold: kill stronghold on key clear
To further prevent cryptographic methods being invoked when the key
has been cleared, we have to unload Stronghold every time the key
is cleared from the memory, as otherwise Stronghold permits such
operation (it only asks for a key to read / write a snapshot).

There are also general code and documentation improvements.