Stronghold Improvements: Timeout and Unloading (#1005)
* stronghold: add timeout and task methods

get_timeout() and set_timeout() are added to view and change the
key clearing timeout. The task will be automatically spawned in
set_timeout() if a new timeout is provided and key is still in the
memory.

restart_key_timeout() is also added as a shortcut of refreshing
the count down by setting the same timeout again.

* secret: add timeout for dto ser/de

* stronghold: no SecretManager methods without key

* Revert "stronghold: no SecretManager methods without key"

This reverts commit 675ae1cd19017e3d29043e7f8b398a881d4d7e2a.

This check has no actual meaning - setting an incorrect key can
allow subsequent cryptography oprations run.

* stronghold: kill stronghold on key clear

To further prevent cryptographic methods being invoked when the key
has been cleared, we have to unload Stronghold every time the key
is cleared from the memory, as otherwise Stronghold permits such
operation (it only asks for a key to read / write a snapshot).

There are also general code and documentation improvements.

* stronghold: increase sleep duration in tests

Poor CI machines sometimes run out of time.

* stronghold: unset snapshot load status on unload

* secret: fix import warning

* stronghold: unload on key clear

* stronghold: fix stronghold unload on key clear

Unloading snapshot requires the key to flush the snapshot onto
disk, so the key must be retained upon unloading.

* secret: fix variable namings

* stronghold,secret: remove unwraps

Remove calls to .unwrap() to avoid panics. As a result, the build
method of StrongholdAdapterBuilder becomes failable and is renamed
to try_build. Existing codes are adjusted accordingly. General code
and documentation improvements are done too.